Move connectors telemetry code into its own modules
