+++ apollo-router/src/plugins/telemetry/config_new/connectors.rs
@@ -1,0 +1,3 @@
+//! Connectors telemetry.
+
+pub(crate) mod http;
