+++ apollo-router/src/plugins/telemetry/config_new/connectors/http.rs
@@ -1,0 +1,5 @@
+//! Telemetry related to Connectors HTTP requests and responses.
+
+pub(crate) mod attributes;
+pub(crate) mod instruments;
+pub(crate) mod selectors;
