+++ apollo-router/src/plugins/telemetry/config_new/mod.rs
@@ -20,6 +20,7 @@ pub(crate) mod conditions;
 
 pub(crate) mod cache;
 mod conditional;
+pub(crate) mod connectors;
 pub(crate) mod cost;
 pub(crate) mod events;
 mod experimental_when_header;
